Cleanup pass + add Addresses.removeAddress (#99)
* remove unused file


* fmt

* checkpoint

* fmt

* fmt

* fmt

* alphabetically ordered

* fmt

* feat: add removeAddress to Addresses registry

Implements removeAddress/resetRemovedAddresses/getRemovedAddresses on
Addresses and IAddresses, mirroring the existing add/change tracking
pattern. This gives callers a way to undo address entries (and the
corresponding entry in the on-disk json via updateJson()).

test_checkAddressFileUpdate now calls removeAddress + updateJson() as
teardown so it no longer permanently writes its TEST1/TEST2/TEST3
fixtures into the committed addresses/*.json files, which was causing
it to fail on repeat runs with "Address with name: TEST1 already set
on chain: 11155111".


* chore: normalize SPDX license headers to MIT

The project's LICENSE and README both declare Forge Proposal Simulator
as MIT licensed, but several files carried GPL-3.0-or-later or
BSD-3-Clause SPDX headers (or none at all). Normalize every non-vendored
.sol file to MIT to match the project's actual, declared license and
make it fully permissive/open source.

Vendored dependencies under lib/ are untouched.


---------

+++ mocks/MockAuction.sol
@@ -1,9 +1,10 @@
+// SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
 /// @notice This is a mock contract for testing purposes only, it SHOULD NOT be used in production.
 contract MockAuction {
     address public highestBidder;
-    uint public highestBid;
+    uint256 public highestBid;
     address public owner;
 
     constructor() {
@@ -18,7 +19,7 @@ contract MockAuction {
         highestBid = msg.value;
 
         if (previousBidder != address(0)) {
-            (bool success, ) = payable(previousBidder).call{value: msg.value}("");
+            (bool success,) = payable(previousBidder).call{value: msg.value}("");
             assert(success);
         }
     }
